updateDoneId

+++ src/main/java/bike/service/app/model/repository/UsersRepository.java
@@ -7,8 +7,8 @@ import bike.service.app.model.Users;
 
 
 @Repository
-public interface UsersRepository extends JpaRepository <Users, Integer> {
+public interface UsersRepository extends JpaRepository <Users, Long> {
     //getting user name from repository tzw konwencja nazewnictwa tak dziala w springBoot
     public Users findByUserName(String userName);
-    public Users findByUserId(int userId);
+    public Users findByUserId(long userId);
 }
